# Conversion performance improvement