--- a/qiskit_optimization/problems/quadratic_expression.py
+++ b/qiskit_optimization/problems/quadratic_expression.py
@@ -17,6 +17,7 @@ from typing import List, Union, Dict, Tuple, Any
 import numpy as np
 from numpy import ndarray
 from scipy.sparse import spmatrix, dok_matrix, tril, triu
+import sys
 
 from .quadratic_program_element import QuadraticProgramElement
 from ..exceptions import QiskitOptimizationError
@@ -105,6 +106,10 @@ class QuadraticExpression(QuadraticProgramElement):
         if isinstance(coefficients, (list, ndarray, spmatrix)):
             coefficients = dok_matrix(coefficients)
         elif isinstance(coefficients, dict):
+
+            # Check if the python version is at least 3.9. See pull request #594 for more details on this check. 
+            new_update_rule = sys.version_info >= (3,9)
+
             n = self.quadratic_program.get_num_vars()
             coeffs = dok_matrix((n, n))
             for (i, j), value in coefficients.items():
@@ -113,12 +118,18 @@ class QuadraticExpression(QuadraticProgramElement):
                 if isinstance(j, str):
                     j = self.quadratic_program.variables_index[j]
 
-                if i > j:
-                    coeffs[j, i] += value
+                if new_update_rule:
+                    if i > j:
+                        coeffs[j, i] += value
+                    else:
+                        coeffs[i, j] += value
                 else:
-                    coeffs[i, j] += value
+                    coeffs[i, j] = value
 
-            return coeffs
+            if new_update_rule:
+                return coeffs
+            else:
+                coefficients = coeffs
         else:
             raise QiskitOptimizationError(f"Unsupported format for coefficients: {coefficients}")
         return self._triangle_matrix(coefficients)